EPBDS-10068 parse MANIFEST.MF and display it on UI for each deployment. Fix NPEs
--HG--
branch : 5.23.x

+++ WSFrontend/org.openl.rules.ruleservice.ws/src/org/openl/rules/ruleservice/rest/RulesDeployerRestController.java
@@ -84,6 +84,9 @@ public class RulesDeployerRestController {
     @Produces("application/zip")
     public Response read(@PathParam("serviceName") final String serviceName) throws Exception {
         OpenLService service = ruleServiceManager.getServiceByName(serviceName);
+        if (service == null) {
+            return Response.status(Status.NOT_FOUND).build();
+        }
         FileItem fileItem = rulesDeployerService.read(service.getServicePath());
         return Response.ok(fileItem.getStream())
             .header("Content-Disposition", "attachment;filename='" + serviceName + ".zip'")
@@ -99,6 +102,9 @@ public class RulesDeployerRestController {
     @Path("/delete/{serviceName}")
     public Response delete(@PathParam("serviceName") final String serviceName) throws Exception {
         OpenLService service = ruleServiceManager.getServiceByName(serviceName);
+        if (service == null) {
+            return Response.status(Status.NOT_FOUND).build();
+        }
         boolean deleted = rulesDeployerService.delete(service.getServicePath());
         return Response.status(deleted ? Response.Status.OK : Status.NOT_FOUND).build();
     }
